Updating the bootstrap directly when proxyProtocol is invoked.
